ci: fix long-standing golangci-lint failures on main
Three issues were blocking every CI run since before #29 and didn't
surface in the usual local workflows:

- agent/internal/enricher/hasher.go: drop redundant uint64(st.Dev)
  conversion (unconvert) — syscall.Stat_t.Dev is already uint64 on
  linux, the cast was legacy noise.
- agent/internal/enricher/{hasher_test.go,hash_wire_test.go}: test
  fixtures were writing 0o755 to satisfy shebang-looking bodies, but
  nothing exec'd the files — they're hashed as bytes. Drop to 0o600
  (gosec G306). No behavior change.
- server/internal/config/config_test.go: the intentional-typo fixture
  string tripped the misspell linter; rename to one it doesn't
  recognise so the suggestion path still gets exercised.

No production behavior change. `make lint` and `make test` clean across
all three modules locally.

+++ agent/internal/enricher/hasher.go
@@ -167,7 +167,7 @@ func statKey(path string) (hashKey, bool) {
 		return hashKey{}, false
 	}
 	return hashKey{
-		dev:   uint64(st.Dev),
+		dev:   st.Dev,
 		inode: st.Ino,
 		mtime: st.Mtim.Nano(),
 	}, true
